requirements.txt
azure-storage-blob>=12.13.0